init pdf vector store
