+++ counter.txt
@@ -1,1 +1,1 @@
-4
+6
